Externalize InvokerPool initialization logic.
This piece of logic clutters the loadbalancer's code for no good reason. We should externalize it.
